WIP! Fns mostly in place. Working on a type class for u/plc.
