+++ plutus-core/plutus-core.cabal
@@ -143,6 +143,9 @@ library
         UntypedPlutusCore.Parser
         UntypedPlutusCore.Rename
         UntypedPlutusCore.Check.Uniques
+        UntypedPlutusCore.Core.Type
+        UntypedPlutusCore.Core.Instance.CBOR
+
 
         Common
         Crypto
@@ -222,7 +225,6 @@ library
         UntypedPlutusCore.Analysis.Definitions
         UntypedPlutusCore.Core
         UntypedPlutusCore.Core.Instance
-        UntypedPlutusCore.Core.Instance.CBOR
         UntypedPlutusCore.Core.Instance.Eq
         UntypedPlutusCore.Core.Instance.Flat
         UntypedPlutusCore.Core.Instance.Pretty
@@ -232,7 +234,6 @@ library
         UntypedPlutusCore.Core.Instance.Pretty.Readable
         UntypedPlutusCore.Core.Instance.Recursive
         UntypedPlutusCore.Core.Plated
-        UntypedPlutusCore.Core.Type
         UntypedPlutusCore.Evaluation.Machine.Cek.CekMachineCosts
         UntypedPlutusCore.Evaluation.Machine.Cek.ExBudgetMode
         UntypedPlutusCore.Evaluation.Machine.Cek.Internal
@@ -412,8 +413,10 @@ test-suite untyped-plutus-core-test
 
 executable plc
     import: lang
-    main-is: Main.hs
-    hs-source-dirs: executables/plc
+    main-is: plc/Main.hs
+    hs-source-dirs: executables
+    other-modules:
+        Common
     build-depends:
         plutus-core -any,
         base <5,
@@ -421,6 +424,7 @@ executable plc
         deepseq -any,
         flat -any,
         monoidal-containers -any,
+        mtl -any,
         optparse-applicative -any,
         prettyprinter -any,
         serialise -any,
@@ -430,8 +434,10 @@ executable plc
 
 executable uplc
     import: lang
-    main-is: Main.hs
-    hs-source-dirs: executables/uplc
+    main-is: uplc/Main.hs
+    hs-source-dirs: executables
+    other-modules:
+        Common
     build-depends:
         plutus-core -any,
         base <5,
@@ -439,6 +445,7 @@ executable uplc
         deepseq -any,
         flat -any,
         monoidal-containers -any,
+        mtl -any,
         optparse-applicative -any,
         prettyprinter -any,
         serialise -any,
